requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
python-dotenv>=1.0
PyMySQL>=1.0
Flask-Login>=0.6
Flask-WTF>=1.0